requirements.txt
pandas 
numpy
scikit-learn
python-dotenv
pymongo
pymongo[srv]==3.6
setuptools
certifi
dotenv
PyYAML
dill

# -e .